+++ src/pages/error/Error.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 function Error() {
   return <div>Error</div>;
 }
