+++ src/pages/main/Main.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import * as S from './Main.styled';
 const Main = () => {
   /*
